added tabs mockup and updated note conente style
